refactor : change packaging
 - 패키징 변경
 - 놓친 테스트 코드 추가
